Added an onPeak callback
Added an onPeak callback that the drawer fires as soon as it decides it will handle a pan.  My use case is that I want to dismiss the keyboard as soon as the drag starts.  You may have ideas on better ways to merge this with other concepts, but it's accomplishing what I was looking for in my project.