Honor Tap To Close with Negotiate Pan
Might want some additional testing but this seems to honor taps within the mask properly.  Previously, if `negotiatePan={true}` the drawer would let the main content respond to taps/gestures instead of capturing them with the drawer when `captureGestures={true}`.

There's some goofy behavior if `captureGestures={true}` but none of the acceptTap/acceptDoubleTap/tapToClose are set to true with  `captureGestures={true}` so some additional work is likely needed.  But this is working for my use case.